requirements.txt
appdirs==1.4.3
cssmin==0.2.0
Django==1.11.1
django-filter==1.0.4
django-pipeline==1.6.12
djangorestframework==3.6.3
Markdown==2.6.8
packaging==16.8
ply==3.10
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
slimit==0.8.1
